Window max height change, scatter+reset buttons added, init hookups, etc
Window max height edited, added a scatter and reset button to bottom of UI with ability to stretch and give a bit of space between the bottom buttons and the rest of the UI. Set default values on some variables in init of ScatterObject. Called scatterobject in ScatterUI class and set up hookups between ui variables and ScatterObject init variables. Also added some test button functionality that will be amended shortly, along with an @QtCore.Slot() def.