Quick if/else check to len to prevent list index out of range traceback.
Adds if/else to def select_scatter_object functionality and checks to see if object has been selected. If not, prevents traceback error and logs warning to Maya. If object was previously selected and then deselected and the button was pressed once more, the old text would remain in the QLineEdit. current_object_def is set to None in the else statement to prevent this and essentially reset the QLineEdit to be blank.