Added functionality for selecting and setting scatter destination
Added hookup to button for selecting scatter destination, set up code that scatters the object as well as the randomization functionality. Set up additional error checking for if nothing is selected in the target destination field (if nothing, warning is given). This also happens if an entire object is put in the target destination instead of specific vertices.