Reworked select scatter def to allow an entire object to be selected
Takes an object that is selected and does polyListComponentConversion and filterExpand. Checks to see if anything was passed into those, if not, logs a warning to tell the user that no target was selected when they pressed the select target button. If verts or object (meaning all verts) are passed, then it sets current_target_def equal to scatter_target_def (which had handled the previous logic). Current target def then updates the QLineEdit text while scatter_target_def is used in the other def that actually does the scatter functionality.