Added functionality to scatter on % of selected vertices, PEP8 fixes
Added functionality for a new UI element, being a QSpinBox for the scatter percentage. Also hooked up functionality for randomly selecting a percentage of a vertices list. Built in log warning that bars the running of the percentage scatter if the percentage is 0, since nothing happens. Informs the user to select a higher percentage.