Additional feature implementation - position embed offset UI and defs.
Adds a position offset to allow users to choose how much they may wish to embed their scattered objects under the surface. Has two versions, one for aligning to normals and aligning to normals + rotation and the other for default settings (AKA no aligning to normals, works as intended, mainly for planes compared to the first). Also implemented UI hookups for this as well and adjusted max height of the GUI itself.